// src/components/FileUploader/types.ts
export enum FileStatus {
  ADDED = 'added',
  QUEUED = 'queued',
  UPLOADING = 'uploading',
  PAUSED = 'paused',
  ERROR = 'error',
  UPLOADED = 'uploaded',
}

export interface StorageFile {
  id: string;
  file?: File;
  status: FileStatus;
  progress: number;
  // only present after an upload completes
  resolvedKey?: string;
  uploadTask?: UploadTask;
  key: string;
  error: string;
  isImage: boolean;
}

export type StorageFiles = StorageFile[];

export type DefaultFile = Pick<StorageFile, 'key'>;

export interface ProcessFileParams extends Record<string, any> {
  file: File;
  key: string;
  useAccelerateEndpoint?: boolean;
}

export type ProcessFile = (
  params: ProcessFileParams
) => Promise<ProcessFileParams> | ProcessFileParams;

//export type ProcessFileError = ProcessFileParams & { error: string };

export interface ProcessFileErrorParams extends Record<string, any> {
  file: File;
  key: string;
  useAccelerateEndpoint?: boolean;
  error: string | Error;
}

export type ProcessFileError = (error: ProcessFileErrorParams) => void;
//) => Promise<ProcessFileError> | ProcessFileError;

export interface FileUploaderHandle { clearFiles: () => void; }

export type StorageAccessLevel = 'guest' | 'protected' | 'private';

export interface FileUploaderProps {
  /**
   * List of accepted File types, values of `['*']` or undefined allow any files
   * @see https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/accept
   */
  acceptedFileTypes?: string[];
  /**
   * Access level for file uploads
   * @see https://docs.amplify.aws/lib/storage/configureaccess/q/platform/js/
   */
  accessLevel: StorageAccessLevel;

  /**
   * Determines if the upload will automatically start after a file is selected, default value: true
   */
  autoUpload?: boolean;
  /**
   * Component overrides
   */
  components?: {
    Container?: React.ComponentType<ContainerProps>;
    DropZone?: React.ComponentType<DropZoneProps>;
    FileList?: React.ComponentType<FileListProps>;
    FilePicker?: React.ComponentType<FilePickerProps>;
    FileListHeader?: React.ComponentType<FileListHeaderProps>;
    FileListFooter?: React.ComponentType<FileListFooterProps>;
  };
  /**
   * List of default files already uploaded
   */
  defaultFiles?: DefaultFile[];
  /**
   * Overrides default display text
   */
  displayText?: FileUploaderDisplayText;
  /**
   * Determines if upload can be paused / resumed
   */
  isResumable?: boolean;
  /**
   * Maximum total files to upload in each batch
   */
  maxFileCount: number;
  /**
   * Maximum file size in bytes
   */
  maxFileSize?: number;
  /**
   * When a file is removed
   */
  onFileRemove?: (file: { key: string }) => void;
  /**
   * Monitor upload errors
   */
  onUploadError?: (error: string, file: { key: string }) => void;
  /**
   * Monitor upload success
   */
  onUploadSuccess?: (event: { key?: string }) => void;
  /**
   * When a file begins uploading
   */
  onUploadStart?: (event: { key?: string }) => void;
  /**
   * Process file before upload
   */
  processFile?: ProcessFile;
  /**
   * Monitor process file errors
   */
  onProcessFileError?: (error: ProcessFileErrorParams) => void;
  /**
   * Determines if thumbnails show for image files
   */
  showThumbnails?: boolean;
  /**
   * Provided value is prefixed to the file `key` for each file
   */
  path?: string;

  useAccelerateEndpoint?: boolean;
}

export interface FileUploaderPathProps
  extends Omit<FileUploaderProps, 'accessLevel' | 'path'> {
  /**
   * S3 bucket key, allows either a `string` or a `PathCallback`:
   * - `string`: `path` is prefixed to the file `key` for each file
   * - `PathCallback`: callback provided an input containing the current `identityId`,
   *    resolved value is prefixed to the file `key` for each file
   */
  path: string | PathCallback;

  accessLevel?: never;

  useAccelerateEndpoint?: boolean;
}